Add `exec` to Robot class
Add an execute method `exec` to Robot, which accepts a textual
representation of known command on Robot and then proceeds to execute
that method when they match; i.e. [robot exec "MOVE"] calls
[robot move]. This will be useful for when processing incoming User
inputs either from the console prompt directly or via a file containg a
list of commands indirectly.

== References:
- [switch manual page \- Tcl Built\-In Commands]
  (https://www.tcl.tk/man/tcl/TclCmd/switch.htm)

- [Tcl Switch Statement]
  (https://www.tutorialspoint.com/tcl-tk/tcl_switch_statement.htm)